+++ requirements.txt
@@ -1,0 +1,8 @@
+# We pin cryptography to 3.4.8 due to a build error that occurs with the
+# current version (36.0.0 at the time of this commit). In short, building
+# the cryptography package with Rust results in a segfault, so we need to
+# set the CRYPTOGRAPHY_DONT_BUILD_RUST env var to force the package to build
+# with C instead. This env var is only present in cryptography<3.5.
+cryptography==41.0
+pyfxa==0.7.7
+tokenlib==2.0.0
