+++ tokenserver-auth/src/token.rs
@@ -1,0 +1,11 @@
+#[cfg(not(feature = "py"))]
+mod native;
+
+#[cfg(feature = "py")]
+mod py;
+
+#[cfg(feature = "py")]
+pub type Tokenlib = py::PyTokenlib;
+
+#[cfg(not(feature = "py"))]
+pub type Tokenlib = native::Tokenlib;
